% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{blue_sequential_wfp}
\alias{blue_sequential_wfp}
\title{Blue sequential palette}
\format{
a named vector of hex codes representing the blue sequential palette.
}
\usage{
blue_sequential_wfp
}
\description{
A sequential color palette based on the main blue color.
}
\keyword{datasets}
