% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{cari_area_palette}
\alias{cari_area_palette}
\title{CARI Area categorical palette}
\format{
a named vector of hex codes representing the CARI area palette.
}
\usage{
cari_area_palette
}
\description{
Palette for CARI area classification with 4 categories.
}
\keyword{datasets}
