% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{cari_palette}
\alias{cari_palette}
\title{CARI sequential palette}
\format{
a named vector of hex codes representing the CARI palette.
}
\usage{
cari_palette
}
\description{
CARI palette
}
\keyword{datasets}
