% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_wfp_all.R
\name{display_wfp_all}
\alias{display_wfp_all}
\title{WFP color palettes viewer}
\usage{
display_wfp_all(n = NULL, type = "all")
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on the palette (2), maximum depending on the palette (5, 7 or 10)}

\item{type}{Type of the palette, can be "sequential", "diverging", "qualitative" or "all"}
}
\description{
Displays the color palettes from wfpthemes package color
}
\examples{
library(wfpthemes)
display_wfp_all()
display_wfp_all(type = "diverging")
display_wfp_all(type = c("diverging", "qualitative", "sequential"))
display_wfp_all(3, type = "sequential")
display_wfp_all(5, type = "sequential")

}
