% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{food_security_blue_3cat}
\alias{food_security_blue_3cat}
\title{Food security categorical blue palette (3 categories)}
\format{
a named vector of hex codes representing the food security blue palette.
}
\usage{
food_security_blue_3cat
}
\description{
Blue categorical palette for food security classification with 3 categories.
}
\keyword{datasets}
