% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{food_security_red_4cat}
\alias{food_security_red_4cat}
\title{Food security categorical red palette (4 categories)}
\format{
a named vector of hex codes representing the food security red palette.
}
\usage{
food_security_red_4cat
}
\description{
Red categorical palette for food security classification with 4 categories.
}
\keyword{datasets}
