% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{ipc_palette}
\alias{ipc_palette}
\title{IPC sequential palette}
\format{
a named vector of hex codes representing the IPC palette.
}
\usage{
ipc_palette
}
\description{
IPC palette
}
\keyword{datasets}
