% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_black}
\alias{main_black}
\title{Main black color}
\format{
A character string of length 1.
}
\usage{
main_black
}
\description{
A hex code representing the black color used in the package.
}
\keyword{datasets}
