% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_blue}
\alias{main_blue}
\title{Main Blue Color}
\format{
A character string of length 1.
}
\usage{
main_blue
}
\description{
A hex code representing the main blue color used in the package.
}
\keyword{datasets}
