% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_ivory}
\alias{main_ivory}
\title{Main ivory color}
\format{
A character string of length 1.
}
\usage{
main_ivory
}
\description{
A hex code representing the ivory color used in the package.
}
\keyword{datasets}
