% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_light_blue}
\alias{main_light_blue}
\title{Main light blue color}
\format{
A character string of length 1.
}
\usage{
main_light_blue
}
\description{
A hex code representing the light blue color used in the package.
}
\keyword{datasets}
