% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_light_red}
\alias{main_light_red}
\title{Main light red color}
\format{
A character string of length 1.
}
\usage{
main_light_red
}
\description{
A hex code representing the light red color used in the package.
}
\keyword{datasets}
