% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_navy}
\alias{main_navy}
\title{Main navy blue color}
\format{
A character string of length 1.
}
\usage{
main_navy
}
\description{
A hex code representing the navy blue color used in the package.
}
\keyword{datasets}
