% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_orange}
\alias{main_orange}
\title{Main orange color}
\format{
A character string of length 1.
}
\usage{
main_orange
}
\description{
A hex code representing the orange color used in the package.
}
\keyword{datasets}
