% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_purple}
\alias{main_purple}
\title{Main orange color}
\format{
A character string of length 1.
}
\usage{
main_purple
}
\description{
A hex code representing the orange color used in the package.
}
\keyword{datasets}
