% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_red}
\alias{main_red}
\title{Main red color}
\format{
A character string of length 1.
}
\usage{
main_red
}
\description{
A hex code representing the red color used in the package.
}
\keyword{datasets}
