% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_tan}
\alias{main_tan}
\title{Main tan color}
\format{
A character string of length 1.
}
\usage{
main_tan
}
\description{
A hex code representing the tan color used in the package.
}
\keyword{datasets}
