% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{main_white}
\alias{main_white}
\title{Main white color}
\format{
A character string of length 1.
}
\usage{
main_white
}
\description{
A hex code representing the white color used in the package.
}
\keyword{datasets}
