% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal_grey_wfp}
\alias{pal_grey_wfp}
\title{Grey sequential palette}
\format{
a named vector of hex codes representing a monocolor sequential palette.
}
\usage{
pal_grey_wfp
}
\description{
A sequential monocolor palette
}
\keyword{datasets}
