% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal_main}
\alias{pal_main}
\title{Main WFP VAM & Monitoring Palette for Food Security Indicators}
\format{
a named vector of hex codes representing the main WFP VAM & Monitoring palette.
}
\usage{
pal_main
}
\description{
A palette containing the main WFP VAM & Monitoring colors
}
\keyword{datasets}
