% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal_navy_wfp}
\alias{pal_navy_wfp}
\title{Navy sequential palette}
\format{
a named vector of hex codes representing the navy sequential palette.
}
\usage{
pal_navy_wfp
}
\description{
A sequential color palette based on the navy color.
}
\keyword{datasets}
