% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{pal_orange_wfp}
\alias{pal_orange_wfp}
\title{Orange sequential palette}
\format{
a named vector of hex codes representing a Orange sequential palette.
}
\usage{
pal_orange_wfp
}
\description{
A sequential orange palette
}
\keyword{datasets}
