% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{red_sequential_wfp}
\alias{red_sequential_wfp}
\title{Red sequential palette}
\format{
a named vector of hex codes representing the red sequential palette.
}
\usage{
red_sequential_wfp
}
\description{
A sequential color palette based on the red color.
}
\keyword{datasets}
