% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sampledataenglish}
\alias{sampledataenglish}
\title{World Food Programme Sample Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3000 rows and 53 columns.
}
\usage{
sampledataenglish
}
\description{
A sample of food security survey data from the World Food Programme
}
\details{
#' @source Generated for demonstration purposes
}
\keyword{datasets}
