% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_wfp_b}
\alias{scale_fill_wfp_b}
\title{scale_fill_wfp_b}
\usage{
scale_fill_wfp_b(
  ...,
  type = "qualitative",
  palette = 1,
  direction = 1,
  nmax = NULL,
  order = NULL,
  na.value = "#E9E9E9"
)
}
\arguments{
\item{...}{Additional arguments passed to `discrete_scale`.}

\item{type}{Character. Type of palette to use ("qualitative", etc.). Default is "qualitative".}

\item{palette}{Numeric or Character. Palette index or name. Default is 1.}

\item{direction}{Numeric. Direction of the palette (1 for normal, -1 for reversed). Default is 1.}

\item{nmax}{Numeric. Maximum number of colors in the palette. Default is NULL.}

\item{order}{Numeric. Order of colors in the palette. Default is NULL.}

\item{na.value}{Character. Color for missing values. Default is "#E9E9E9".}
}
\value{
A ggplot2 discrete fill scale.
}
\description{
Creates a discrete fill scale for ggplot2 visualizations.
}
