% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wfp.R
\name{theme_wfp}
\alias{theme_wfp}
\title{theme_wfp}
\usage{
theme_wfp(
  font_size = 10,
  font_family = "Open Sans",
  line_size = 0.5,
  rel_tiny = 6/10,
  rel_small = 8/10,
  rel_normal = 10/10,
  rel_large = 12/10,
  rel_xlarge = 14/10,
  rel_xxlarge = 16/10,
  grid = TRUE,
  axis = "x",
  axis_text = TRUE,
  axis_title = TRUE,
  axis_ticks = FALSE,
  legend = TRUE,
  legend_title = FALSE
)
}
\arguments{
\item{font_size}{Numeric. Base font size for the theme. Default is 10.}

\item{font_family}{Character. Font family to use. Default is "Open Sans".}

\item{line_size}{Numeric. Line size for elements like grid lines and axis lines. Default is 0.5.}

\item{rel_tiny}{Numeric. Relative size for tiny text. Default is 6/10.}

\item{rel_small}{Numeric. Relative size for small text. Default is 8/10.}

\item{rel_normal}{Numeric. Relative size for normal text. Default is 10/10.}

\item{rel_large}{Numeric. Relative size for large text. Default is 12/10.}

\item{rel_xlarge}{Numeric. Relative size for extra-large text. Default is 14/10.}

\item{rel_xxlarge}{Numeric. Relative size for extra-extra-large text. Default is 16/10.}

\item{grid}{Logical or Character. Whether to display grid lines. Can also specify "x", "y", "X", or "Y" for specific grid lines. Default is TRUE.}

\item{axis}{Logical or Character. Whether to display axis lines. Can also specify "x" or "y" for specific axes. Default is "x".}

\item{axis_text}{Logical or Character. Whether to display axis text. Can also specify "x" or "y" for specific axes. Default is TRUE.}

\item{axis_title}{Logical or Character. Whether to display axis titles. Can also specify "x" or "y" for specific axes. Default is TRUE.}

\item{axis_ticks}{Logical or Character. Whether to display axis ticks. Can also specify "x" or "y" for specific axes. Default is FALSE.}

\item{legend}{Logical. Whether to display the legend. Default is TRUE.}

\item{legend_title}{Logical. Whether to display the legend title. Default is FALSE.}
}
\value{
A ggplot2 theme object customized for WFP visualizations.
}
\description{
A custom ggplot2 theme function designed for WFP visualizations.
}
