% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wfp.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{update_geom_font_defaults}
\usage{
update_geom_font_defaults(
  family = "Open Sans",
  face = "plain",
  size = 3.5,
  color = "#191919"
)
}
\arguments{
\item{family}{Character. Font family to use. Default is "Open Sans".}

\item{face}{Character. Font face to use (e.g., "plain", "bold"). Default is "plain".}

\item{size}{Numeric. Font size to use. Default is 3.5.}

\item{color}{Character. Font color to use. Default is "#191919" (dark grey).}
}
\value{
Updates the default settings for text and label geoms in ggplot2.
}
\description{
Updates the default font family, face, size, and color for text and label geoms in ggplot2.
}
