% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{wfp_pal}
\alias{wfp_pal}
\title{wfp_pal}
\usage{
wfp_pal(n = NULL, name, ...)
}
\arguments{
\item{n}{Numeric. Number of colors to generate. Default is NULL.}

\item{name}{Character. Name of the palette.}

\item{...}{Additional arguments passed to `colorRampPalette`.}
}
\value{
A vector of colors.
}
\description{
Validates the number of colors in a palette and generates the palette.
}
\examples{
wfp_pal(5, "blue_sequential")
}
