% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{wfp_pal_name}
\alias{wfp_pal_name}
\title{wfp_pal_name}
\usage{
wfp_pal_name(palette, type)
}
\arguments{
\item{palette}{Character. Palette name to validate.}

\item{type}{Character. Type of palette ("qualitative", "sequential", etc.).}
}
\value{
A valid palette name.
}
\description{
Validates the palette name and replaces invalid names with "blue_sequential".
}
\examples{
wfp_pal_name("invalid_palette", "qualitative")
}
