% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{wfp_pal_scale}
\alias{wfp_pal_scale}
\title{wfp_pal_scale}
\usage{
wfp_pal_scale(
  type = "qualitative",
  nmax = NULL,
  order = NULL,
  palette = 1,
  direction = 1
)
}
\arguments{
\item{type}{Character. Type of palette to use ("qualitative", "sequential", etc.). Default is "qualitative".}

\item{nmax}{Numeric. Maximum number of colors in the palette. Default is NULL.}

\item{order}{Numeric. Order of colors in the palette. Default is NULL.}

\item{palette}{Numeric or Character. Palette index or name. Default is 1.}

\item{direction}{Numeric. Direction of the palette (1 for normal, -1 for reversed). Default is 1.}
}
\value{
A function that generates a palette for ggplot2 scales.
}
\description{
Creates a palette function for ggplot2 scales.
}
