% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{display_wfp_all}
\alias{display_wfp_all}
\title{display_wfp_all}
\usage{
display_wfp_all(n = NULL, type = "all")
}
\arguments{
\item{n}{Numeric. Number of colors to display for each palette. Default is NULL.}

\item{type}{Character. Type of palettes to display ("all", "qualitative", "sequential", "diverging"). Default is "all".}
}
\value{
A plot displaying the palettes.
}
\description{
Displays all WFP color palettes.
}
\examples{
display_wfp_all()
}
